src/requirements.txt
streamlit==1.42.2
google-cloud-bigquery
google-cloud-storage
langchain
langchain-community
python-dotenv 
plotly
db-dtypes
pydantic-ai
detoxify
sentence-transformers
mlflow>=2.3.0
sqlparse>=0.4.3
chromadb
pandas
numpy
thefuzz
fuzzywuzzy
pytest
coverage
langchain_google_vertexai
scikit-learn
matplotlib
seaborn
shap
xgboost
flask==2.3.3
flask-cors==4.0.0